requirements.txt
streamlit
pillow
numpy
scikit-learn
colorgram.py
opencv-python
reportlab
fpdf
